python/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
asyncio==3.4.3
attrs==24.2.0
av==13.0.0
backoff==2.2.1
base58==2.1.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.4
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.8
command-not-found==0.3
cryptography==3.4.8
dataclasses-json==0.6.7
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
docker==7.1.0
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
emoji==2.14.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
filetype==1.2.0
frozenlist==1.5.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.69.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.6
httplib2==0.20.2
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
importlib-metadata==4.6.4
jeepney==0.7.1
jiter==0.7.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==3.2.0
keyring==23.5.0
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-text-splitters==0.3.5
langdetect==1.0.9
langsmith==0.2.10
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==5.3.0
lz4==4.3.3
marshmallow==3.25.0
more-itertools==8.10.0
multidict==6.1.0
mypy-extensions==1.0.0
ndjson==0.3.1
nest-asyncio==1.6.0
netifaces==0.11.0
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.0
olefile==0.47
openai==1.54.4
opencv-contrib-python==4.10.0.84
orjson==3.10.14
packaging==24.2
paramiko==3.5.0
picamera2==0.3.18
pidng==4.0.9
piexif==1.1.3
pillow==10.4.0
pinecone==5.3.1
pinecone-client==5.0.1
pinecone-plugin-assistant==1.0.0
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
propcache==0.2.1
protobuf==4.25.5
protoc-gen-openapiv2==0.0.1
psutil==6.1.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.7.1
pydantic_core==2.23.4
PyGObject==3.42.1
PyJWT==2.3.0
PyNaCl==1.5.0
pyparsing==2.4.7
pypdf==5.1.0
pyrsistent==0.20.0
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
python-oxmsg==0.0.1
python-prctl==1.8.1
PyYAML==5.4.1
RapidFuzz==3.11.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
RPi.GPIO==0.7.1
SecretStorage==3.3.1
simplejpeg==1.7.6
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
systemd-python==234
tenacity==9.0.0
texttable==1.7.0
tqdm==4.67.0
typing-inspect==0.9.0
typing_extensions==4.12.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
unstructured==0.16.13
unstructured-client==0.28.1
urllib3==2.2.3
v4l2-python3==0.3.5
wadllib==1.3.6
webencodings==0.5.1
websocket-client==0.59.0
wrapt==1.17.1
yarl==1.18.3
zipp==1.0.0
